core: Moved NullRenderer into its own module in preparation for refactoring
